wireless-security: prepare text entries according to 'hints'
Use the hints passed down from the secret agent request, if any, to
select the EAPMethodSimple UI and show or hide GtkEntry widgets
according to which secrets are being requested.  This unfortunately
touches all the places that use EAPMethodSimple.  I used this specific
class because it is already kind of handling a bunch of different
actual EAP methods but I could also have created a new UI separate
EAPMethodSimple.

Since I'm abusing EAPMethodSimple for this, I also needed to add a new
eap method type "unknown", first because where this code is useful NM
does not know what specific EAP method is used and secondly the UI
doesn't really need to know that.

Note that hints being non-NULL implies secrets_only.  Since with
secrets_only the password storage flags can't be changed, I didn't
bother adding the storage flags to the GtkEntry for the private key
passphrase so it doesn't get the secondary icon thing even though the
plain password field does still have the icon.  Only the three hint
strings are supported which are actually in use in the NM tree in the
IWD wifi backend.