wireless-security: show no settings for method "external" in editor mode
The "external" EAP method string is used in connections automatically
created by NM to mirror IWD-configured networks where we only know
they use EAP authentication but not what method or other 802-1x
setting values, and the values are not even needed by NM.  So still
don't allow manually setting a connection's EAP method to "external" but
if that value is already set for the connection then reflect that in the
security combo.  We abuse the EAPMethodSimple class to produce an empty
GtkGrid basically.