wifi-security: Use NMAWs classes instead of WirelessSecurity
This is an almost mechanical conversion to use the NMAWs classes and
remove the dependency on wireless-security to work around issue #92.
There may be things in the libnma version that are not taken advantage
of in this patch, it's just the conversion.

This is still leaking memory as much as it was before.