8021x-security: Use NMAWs classes instead of WirelessSecurity
Convert the ethernet 8021x-security page to use the NMAWs8021x class
and remove the dependency on wireless-security to work around issue #92.
Take advantage of libnma having a pure 8021x ui (without the WPA-EAP
assumptions) meaning that the hack that adds the wireless setting is no
longer needed.

This is still leaking memory as much as it was before.